Remove unused feature support level code
