+++ libs/locales/lib/en/translation.json
@@ -115,7 +115,6 @@
   "ai:Arbiter": "Arbiter",
   "ai:Are you sure you want to reset the cluster?": "Are you sure you want to reset the cluster?",
   "ai:arm64": "arm64",
-  "ai:arm64 is not supported in this OpenShift version": "arm64 is not supported in this OpenShift version",
   "ai:At least 3 hosts are required, capable of functioning as control plane nodes.": "At least 3 hosts are required that are capable of functioning as control plane nodes.",
   "ai:At least one config map is required": "At least one config map is required",
   "ai:At least one of the HTTP or HTTPS proxy URLs is required.": "At least one of the HTTP or HTTPS proxy URLs is required.",
@@ -812,7 +811,6 @@
   "ai:Select one or multiple locations to choose the hosts from.": "Select one or multiple locations of the hosts.",
   "ai:Select this when your hosts are using only IPv4.": "Select this when your hosts are using only IPv4.",
   "ai:selected": "selected",
-  "ai:Selected image does not support arm64": "Selected image does not support arm64",
   "ai:Self Node Remediation requirements": "Self Node Remediation requirements",
   "ai:Serial": "Serial",
   "ai:Serial number": "Serial number",
@@ -831,7 +829,6 @@
   "ai:Single replica": "Single replica",
   "ai:Single replica means components are not expected to be resilient to problems across most fault boundaries associated with high availability. This usually means running critical workloads with just 1 replica and with toleration of full disruption of the component.": "Single replica means components are not expected to be resilient to problems across most fault boundaries associated with high availability. This usually means running critical workloads with just 1 replica and with toleration of full disruption of the component.",
   "ai:Single stack": "Single stack",
-  "ai:Single-Node OpenShift is not supported in this OpenShift version": "Single-Node OpenShift is not supported in this OpenShift version",
   "ai:Size": "Size",
   "ai:SNO: One host is required with at least {{sno_cpu_cores}} CPU cores, {{snoRam}} of RAM, and {{sno_disksize}} GB of disk size storage.": "SNO: One host is required with at least {{sno_cpu_cores}} CPU cores, {{snoRam}} of RAM, and {{sno_disksize}} GB of disk size storage.",
   "ai:Software-Defined Networking (SDN)": "Software-Defined Networking (SDN)",
@@ -990,7 +987,6 @@
   "ai:Use when you have an iPXE server that has already been set up": "Use when you have an iPXE server that has already been set up",
   "ai:useAlerts must be used within AlertsContextProvider": "useAlerts must be used within AlertsContextProvider",
   "ai:Used to describe hosts' physical location. Helps for quicker host selection during cluster creation.": "Used to describe hosts' physical location. Helps for quicker host selection during cluster creation.",
-  "ai:useFeatureSupportLevel must be used within FeatureSupportLevelContextProvider.": "useFeatureSupportLevel must be used within FeatureSupportLevelContextProvider.",
   "ai:User-managed networking": "User-managed networking",
   "ai:Username": "Username",
   "ai:UUID": "UUID",
