+++ libs/ui-lib/lib/cim/components/ClusterDeployment/ClusterDeploymentWizard.tsx
@@ -3,7 +3,6 @@ import { useTranslation } from '../../../common/hooks/use-translation-wrapper';
 import { Grid, GridItem, Wizard, WizardStep } from '@patternfly/react-core';
 import { AlertsContextProvider, ErrorState, LoadingState } from '../../../common';
 import { ClusterDeploymentWizardProps } from './types';
-import { ACMFeatureSupportLevelProvider } from '../featureSupportLevels';
 import { YamlPreview, useYamlPreview } from '../YamlPreview';
 import { wizardStepNames } from './constants';
 import { ClusterDeploymentWizardContextProvider } from './ClusterDeploymentWizardContext';
@@ -72,110 +71,103 @@ export const ClusterDeploymentWizard = ({
     <Grid style={{ height: '100%' }}>
       <GridItem span={isPreviewOpen ? 7 : 12}>
         <AlertsContextProvider>
-          <ACMFeatureSupportLevelProvider clusterImages={clusterImages} isEditClusterFlow={true}>
-            <ClusterDeploymentWizardContextProvider agentClusterInstall={agentClusterInstall}>
-              <Wizard
-                isVisitRequired
-                startIndex={startIndex}
-                className={className}
-                onClose={onClose}
-              >
-                <WizardStep
-                  name={stepNames['installation-type']}
-                  id={'installation-type'}
-                  isDisabled
+          <ClusterDeploymentWizardContextProvider agentClusterInstall={agentClusterInstall}>
+            <Wizard isVisitRequired startIndex={startIndex} className={className} onClose={onClose}>
+              <WizardStep
+                name={stepNames['installation-type']}
+                id={'installation-type'}
+                isDisabled
+              />
+
+              <WizardStep name={stepNames['cluster-details']} id={'cluster-details'}>
+                <ClusterDeploymentDetailsStep
+                  clusterImages={clusterImages}
+                  usedClusterNames={usedClusterNames}
+                  clusterDeployment={clusterDeployment}
+                  agentClusterInstall={agentClusterInstall}
+                  agents={agents}
+                  onSaveDetails={onSaveDetails}
+                  isPreviewOpen={isPreviewOpen}
+                  infraEnv={infraEnv}
+                  isNutanix={isNutanix}
                 />
+              </WizardStep>
 
-                <WizardStep name={stepNames['cluster-details']} id={'cluster-details'}>
-                  <ClusterDeploymentDetailsStep
-                    clusterImages={clusterImages}
-                    usedClusterNames={usedClusterNames}
-                    clusterDeployment={clusterDeployment}
-                    agentClusterInstall={agentClusterInstall}
-                    agents={agents}
-                    onSaveDetails={onSaveDetails}
-                    isPreviewOpen={isPreviewOpen}
-                    infraEnv={infraEnv}
-                    isNutanix={isNutanix}
-                  />
+              <WizardStep name={stepNames['automation']} id={'automation'} isDisabled />
+
+              {isCIMFlow(clusterDeployment) ? (
+                <WizardStep name={stepNames['hosts-selection']} id={'host-selection'}>
+                  {agentClusterInstall?.metadata?.name ? (
+                    <ClusterDeploymentHostSelectionStep
+                      clusterDeployment={clusterDeployment}
+                      agentClusterInstall={agentClusterInstall}
+                      onSaveHostsSelection={onSaveHostsSelection}
+                      agents={agents}
+                      aiConfigMap={aiConfigMap}
+                      onEditRole={hostActions.onEditRole}
+                      onSetInstallationDiskId={hostActions.onSetInstallationDiskId}
+                      isNutanix={isNutanix}
+                    />
+                  ) : (
+                    <LoadingState />
+                  )}
                 </WizardStep>
-
-                <WizardStep name={stepNames['automation']} id={'automation'} isDisabled />
-
-                {isCIMFlow(clusterDeployment) ? (
-                  <WizardStep name={stepNames['hosts-selection']} id={'host-selection'}>
-                    {agentClusterInstall?.metadata?.name ? (
-                      <ClusterDeploymentHostSelectionStep
-                        clusterDeployment={clusterDeployment}
-                        agentClusterInstall={agentClusterInstall}
-                        onSaveHostsSelection={onSaveHostsSelection}
-                        agents={agents}
-                        aiConfigMap={aiConfigMap}
-                        onEditRole={hostActions.onEditRole}
-                        onSetInstallationDiskId={hostActions.onSetInstallationDiskId}
-                        isNutanix={isNutanix}
-                      />
-                    ) : (
-                      <LoadingState />
-                    )}
-                  </WizardStep>
-                ) : (
-                  <WizardStep name={stepNames['hosts-discovery']} id={'host-discovery'}>
-                    {isAIFlow && onSaveISOParams && onCreateBMH ? (
-                      <ClusterDeploymentHostsDiscoveryStep
-                        clusterDeployment={clusterDeployment}
-                        agentClusterInstall={agentClusterInstall}
-                        agents={agents}
-                        bareMetalHosts={bareMetalHosts}
-                        aiConfigMap={aiConfigMap}
-                        infraEnv={infraEnv}
-                        fetchSecret={fetchSecret}
-                        onChangeBMHHostname={onChangeBMHHostname}
-                        onEditRole={hostActions.onEditRole}
-                        onSetInstallationDiskId={hostActions.onSetInstallationDiskId}
-                        onDeleteHost={hostActions.onDeleteHost}
-                        onSaveISOParams={onSaveISOParams}
-                        onCreateBMH={onCreateBMH}
-                        {...rest}
-                      />
-                    ) : (
-                      <LoadingState />
-                    )}
-                  </WizardStep>
-                )}
-
-                <WizardStep name={stepNames['networking']} id={'networking'}>
-                  <ClusterDeploymentNetworkingStep
-                    clusterDeployment={clusterDeployment}
-                    agentClusterInstall={agentClusterInstall}
-                    agents={agents}
-                    onSaveNetworking={onSaveNetworking}
-                    onEditHost={hostActions.onEditHost}
-                    onEditRole={hostActions.onEditRole}
-                    fetchInfraEnv={fetchInfraEnv}
-                    isPreviewOpen={isPreviewOpen}
-                    onSetInstallationDiskId={hostActions.onSetInstallationDiskId}
-                    isNutanix={isNutanix}
-                  />
+              ) : (
+                <WizardStep name={stepNames['hosts-discovery']} id={'host-discovery'}>
+                  {isAIFlow && onSaveISOParams && onCreateBMH ? (
+                    <ClusterDeploymentHostsDiscoveryStep
+                      clusterDeployment={clusterDeployment}
+                      agentClusterInstall={agentClusterInstall}
+                      agents={agents}
+                      bareMetalHosts={bareMetalHosts}
+                      aiConfigMap={aiConfigMap}
+                      infraEnv={infraEnv}
+                      fetchSecret={fetchSecret}
+                      onChangeBMHHostname={onChangeBMHHostname}
+                      onEditRole={hostActions.onEditRole}
+                      onSetInstallationDiskId={hostActions.onSetInstallationDiskId}
+                      onDeleteHost={hostActions.onDeleteHost}
+                      onSaveISOParams={onSaveISOParams}
+                      onCreateBMH={onCreateBMH}
+                      {...rest}
+                    />
+                  ) : (
+                    <LoadingState />
+                  )}
                 </WizardStep>
+              )}
 
-                <WizardStep name={stepNames['custom-manifests']} id={'custom-manifests'}>
-                  <ClusterDeploymentCustomManifestsStep agentClusterInstall={agentClusterInstall} />
-                </WizardStep>
+              <WizardStep name={stepNames['networking']} id={'networking'}>
+                <ClusterDeploymentNetworkingStep
+                  clusterDeployment={clusterDeployment}
+                  agentClusterInstall={agentClusterInstall}
+                  agents={agents}
+                  onSaveNetworking={onSaveNetworking}
+                  onEditHost={hostActions.onEditHost}
+                  onEditRole={hostActions.onEditRole}
+                  fetchInfraEnv={fetchInfraEnv}
+                  isPreviewOpen={isPreviewOpen}
+                  onSetInstallationDiskId={hostActions.onSetInstallationDiskId}
+                  isNutanix={isNutanix}
+                />
+              </WizardStep>
 
-                <WizardStep name={stepNames['review']} id={'review'}>
-                  <ClusterDeploymentReviewStep
-                    onFinish={onFinish}
-                    clusterDeployment={clusterDeployment}
-                    agentClusterInstall={agentClusterInstall}
-                    agents={agents}
-                    clusterImages={clusterImages}
-                    infraEnv={infraEnv}
-                  />
-                </WizardStep>
-              </Wizard>
-            </ClusterDeploymentWizardContextProvider>
-          </ACMFeatureSupportLevelProvider>
+              <WizardStep name={stepNames['custom-manifests']} id={'custom-manifests'}>
+                <ClusterDeploymentCustomManifestsStep agentClusterInstall={agentClusterInstall} />
+              </WizardStep>
+
+              <WizardStep name={stepNames['review']} id={'review'}>
+                <ClusterDeploymentReviewStep
+                  onFinish={onFinish}
+                  clusterDeployment={clusterDeployment}
+                  agentClusterInstall={agentClusterInstall}
+                  agents={agents}
+                  clusterImages={clusterImages}
+                  infraEnv={infraEnv}
+                />
+              </WizardStep>
+            </Wizard>
+          </ClusterDeploymentWizardContextProvider>
         </AlertsContextProvider>
       </GridItem>
       {isPreviewOpen && (
