--- libs/ui-lib/lib/cim/components/featureSupportLevels/index.ts
@@ -1,1 +1,0 @@
-export { default as ACMFeatureSupportLevelProvider } from './ACMFeatureSupportLevelProvider';
