+++ libs/ui-lib/lib/cim/components/index.ts
@@ -4,5 +4,4 @@ export * from './InfraEnv';
 export * from './ClusterDeployment';
 export * from './modals';
 export * from './Agent';
-export * from './featureSupportLevels';
 export * from './Hypershift';
