+++ libs/ui-lib/lib/cim/index.ts
@@ -10,7 +10,6 @@ export * as Reducers from '../common/reducers';
 export * from '../common/api';
 export * from '../common/utils';
 export * from '../common/types';
-export * from '../common/features';
 
 export * from '../common/components/ui';
 export * from '../common/components/AlertsContextProvider';
