--- libs/ui-lib/lib/common/components/featureSupportLevels/index.ts
@@ -1,2 +1,0 @@
-export { default as FeatureSupportLevelBadge } from './FeatureSupportLevelBadge';
-export * from './FeatureSupportLevelContext';
