--- libs/ui-lib/lib/common/components/featureSupportLevels/types.ts
@@ -1,12 +1,0 @@
-import {
-  ArchitectureSupportLevelMap,
-  NewFeatureSupportLevelMap,
-} from '../newFeatureSupportLevels/types';
-
-export interface ArchitecturesSupportsLevel {
-  architectures: ArchitectureSupportLevelMap;
-}
-
-export interface FeaturesSupportsLevel {
-  features: NewFeatureSupportLevelMap;
-}
