+++ libs/ui-lib/lib/common/components/index.ts
@@ -8,4 +8,4 @@ export * from './AlertsContextProvider';
 export * from './hosts';
 export * from './fetching';
 export * from './AddHosts';
-export * from './featureSupportLevels';
+export * from './newFeatureSupportLevels';
