+++ libs/ui-lib/lib/ocm/components/featureSupportLevels/index.ts
@@ -1,3 +1,2 @@
 export { default as NewFeatureSupportLevelProvider } from './FeatureSupportLevelProvider';
 export { default as ClusterFeatureSupportLevelsDetailItem } from './ReviewClusterFeatureSupportLevels';
-export * from '../../../common/components/featureSupportLevels/types';
